// frontend/src/api/models/SearchResultDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { FileMetadataDTO } from './FileMetadataDTO';
import {
    FileMetadataDTOFromJSON,
    FileMetadataDTOFromJSONTyped,
    FileMetadataDTOToJSON,
} from './FileMetadataDTO';

/**
 * 
 * @export
 * @interface SearchResultDTO
 */
export interface SearchResultDTO {
    /**
     * 
     * @type {FileMetadataDTO}
     * @memberof SearchResultDTO
     */
    file: FileMetadataDTO;
    /**
     * 
     * @type {number}
     * @memberof SearchResultDTO
     */
    denseScore: number;
    /**
     * 
     * @type {number}
     * @memberof SearchResultDTO
     */
    lexicalScore: number;
    /**
     * 
     * @type {number}
     * @memberof SearchResultDTO
     */
    totalScore: number;
}

/**
 * Check if a given object implements the SearchResultDTO interface.
 */
export function instanceOfSearchResultDTO(value: object): value is SearchResultDTO {
    if (!('file' in value) || value['file'] === undefined) return false;
    if (!('denseScore' in value) || value['denseScore'] === undefined) return false;
    if (!('lexicalScore' in value) || value['lexicalScore'] === undefined) return false;
    if (!('totalScore' in value) || value['totalScore'] === undefined) return false;
    return true;
}

export function SearchResultDTOFromJSON(json: any): SearchResultDTO {
    return SearchResultDTOFromJSONTyped(json, false);
}

export function SearchResultDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): SearchResultDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'file': FileMetadataDTOFromJSON(json['file']),
        'denseScore': json['dense_score'],
        'lexicalScore': json['lexical_score'],
        'totalScore': json['total_score'],
    };
}

export function SearchResultDTOToJSON(value?: SearchResultDTO | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'file': FileMetadataDTOToJSON(value['file']),
        'dense_score': value['denseScore'],
        'lexical_score': value['lexicalScore'],
        'total_score': value['totalScore'],
    };
}